requirements.txt.txt
streamlit
torch
unsloth
transformers
datasets
trl
pytesseract
Pillow
PyMuPDF
numpy
logging
tempfile
